extract node 3d to core pipeline
