+++ crates/bevy_core_pipeline/src/lib.rs
@@ -5,6 +5,7 @@ pub mod core_3d;
 pub mod fullscreen_vertex_shader;
 pub mod fxaa;
 pub mod gizmo_2d;
+pub mod gizmo_3d;
 pub mod prepass;
 pub mod tonemapping;
 pub mod upscaling;
